Updated tqdm to 4.48.0, to fix leaked semaphore issue under Python 3.8
