+++ conda/_vendor/tqdm/LICENSE
@@ -6,18 +6,24 @@ for their respective work, and release the work under the MIT licence
 Exceptions or notable authors are listed below
 in reverse chronological order:
 
-* MPLv2.0 (text below) 2015-2016 (c) Casper da Costa-Luis
+* files: *
+  MPLv2.0 2015-2020 (c) Casper da Costa-Luis
   [casperdcl](https://github.com/casperdcl).
-* MIT 2016 (c) [PR #96] on behalf of Google Inc.
+* files: tqdm/_tqdm.py
+  MIT 2016 (c) [PR #96] on behalf of Google Inc.
+* files: tqdm/_tqdm.py setup.py README.rst MANIFEST.in .gitignore
+  MIT 2013 (c) Noam Yorav-Raphael, original author.
 
-  [PR #96]: https://github.com/tqdm/tqdm/pull/96
+[PR #96]: https://github.com/tqdm/tqdm/pull/96
 
 
 Mozilla Public Licence (MPL) v. 2.0 - Exhibit A
 -----------------------------------------------
 
-This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
-If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
+This Source Code Form is subject to the terms of the
+Mozilla Public License, v. 2.0.
+If a copy of the MPL was not distributed with this file,
+You can obtain one at https://mozilla.org/MPL/2.0/.
 
 
 MIT License (MIT)
