+++ conda/_vendor/tqdm/_version.py
@@ -5,7 +5,7 @@ from io import open as io_open
 __all__ = ["__version__"]
 
 # major, minor, patch, -extra
-version_info = 4, 22, 0
+version_info = 4, 48, 0
 
 # Nice string for the version
 __version__ = '.'.join(map(str, version_info))
